+++ test-packages/test-services-odata-v4/multiple-schemas-service/BatchRequest.ts
@@ -12,6 +12,7 @@ import {
   ODataBatchRequestBuilder,
   UpdateRequestBuilder,
   FunctionImportRequestBuilder,
+  ActionImportRequestBuilder,
   BatchChangeSet
 } from '@sap-cloud-sdk/odata-v4';
 import { transformVariadicArgumentToArray } from '@sap-cloud-sdk/util';
@@ -21,7 +22,11 @@ import {
   TestEntity3,
   TestEntity4,
   TestFunctionImportEntityReturnType1Parameters,
-  TestFunctionImportEntityReturnType2Parameters
+  TestFunctionImportEntityReturnType2Parameters,
+  TestActionImportNoParameterComplexReturnType1Parameters,
+  TestActionImportNoParameterComplexReturnType2Parameters,
+  TestComplexType1,
+  TestComplexType2
 } from './index';
 
 /**
@@ -119,4 +124,14 @@ export type WriteMultipleSchemasServiceRequestBuilder<
   | DeleteRequestBuilder<TestEntity3<DeSerializersT>, DeSerializersT>
   | CreateRequestBuilder<TestEntity4<DeSerializersT>, DeSerializersT>
   | UpdateRequestBuilder<TestEntity4<DeSerializersT>, DeSerializersT>
-  | DeleteRequestBuilder<TestEntity4<DeSerializersT>, DeSerializersT>;
+  | DeleteRequestBuilder<TestEntity4<DeSerializersT>, DeSerializersT>
+  | ActionImportRequestBuilder<
+      DeSerializersT,
+      TestActionImportNoParameterComplexReturnType1Parameters<DeSerializersT>,
+      TestComplexType1
+    >
+  | ActionImportRequestBuilder<
+      DeSerializersT,
+      TestActionImportNoParameterComplexReturnType2Parameters<DeSerializersT>,
+      TestComplexType2
+    >;