+++ test-packages/test-services-odata-common/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @sap-cloud-sdk/test-services-odata-common
 
+## 2.8.0
+
 ## 2.7.1
 
 ## 2.7.0